requirements.txt
jsonschema>=4.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
